classTask_MongoDB_SQLite/PDFmerger&DataStructure/scrapping task perform by fsds students/ineuron scrapping/Bhanu/requirements.txt
alembic==1.7.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
googlesearch-python==1.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==2.6
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
python-decouple==3.6
pytz==2021.3
requests==2.25.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
WTForms==3.0.1
zipp==3.8.0
